remove instances of DS.Store from factories
When using multiple stores with Ember Data, we want to avoid storing the
state on the factory itself. Because multiple stores currently share the
same factory, this means that the first store to look up the factory
will store itself as the `store` on that factory. That means that any
stores thereafter will receive factories whose `store` property
references the first store, not their own.